Change to how transparency is handled
Moved the logic to compute transparency out to a completely separate function for readability, and some optimization benefit in this case.
Also changed the way compound scene objects handle their transparency to make the internal boundaries less harsh. This doesn't completely fix the issue, but it is less noticeable now.